+++ R/geom_raster_filled.R
@@ -1,10 +1,9 @@
 #' geom_raster_filled
-#'
 #' @inheritParams ggplot2::geom_raster
 #' @inheritParams ggplot2::geom_contour
 #' @export
 geom_raster_filled <- function(mapping = NULL, data = NULL,
-                               stat = "raster_filled", position = "identity",
+                               stat = "raster_levels", position = "identity",
                                ...,
                                breaks = NULL,
                                hjust = 0.5,
@@ -13,8 +12,6 @@ geom_raster_filled <- function(mapping = NULL, data = NULL,
                                na.rm = FALSE,
                                show.legend = NA,
                                inherit.aes = TRUE) {
-  # check_number_decimal(hjust)
-  # check_number_decimal(vjust)
   layer(
     data = data,
     mapping = mapping,
@@ -36,7 +33,7 @@ geom_raster_filled <- function(mapping = NULL, data = NULL,
 
 #' @rdname geom_raster_filled
 #' @export
-stat_raster_filled <- function(mapping = NULL, data = NULL,
+stat_raster_levels <- function(mapping = NULL, data = NULL,
                                geom = "raster", position = "identity",
                                ...,
                                # bins = NULL,
@@ -48,7 +45,7 @@ stat_raster_filled <- function(mapping = NULL, data = NULL,
   layer(
     data = data,
     mapping = mapping,
-    stat = StatRasterFilled,
+    stat = StatRasterLevels,
     geom = geom,
     position = position,
     show.legend = show.legend,
@@ -62,26 +59,3 @@ stat_raster_filled <- function(mapping = NULL, data = NULL,
     )
   )
 }
-
-#' @rdname ggplot2-ggproto
-#' @format NULL
-#' @usage NULL
-#' @export
-StatRasterFilled <- ggproto("StatRasterFilled", StatIdentity,
-  required_aes = c("x", "y", "z"),
-  default_aes = aes(fill = after_stat(level)), # order = after_stat(level),
-  # z and weight get dropped during statistical transformation
-  dropped_aes = c("z"),
-  setup_params = function(data, params) {
-    if (is.null(params$breaks)) {
-      params$breaks <- pretty(data$z, 7)
-    }
-    params
-  },
-  setup_data = function(data, params) {
-    data %>% mutate(level = cut(z, params$breaks))
-  },
-  compute_group = function(data, scales, breaks = NULL) {
-    data
-  }
-)